Re-type ComputeY as a List.
The number of unique elements in different columns in x1 may be
different, which leads to attempting to add vectors of different length
into res. Previously, `push_back` on a DataFrame changed the DataFrame
to a list, which is being addressed in RcppCore/Rcpp#1099. This change,
combined with attempting to add vectors of different lengths, leads to
an error.